CurrencyConverter was changed and added new dtos

+++ src/main/java/ua/bank/bankservice/service/TransactionService.java
@@ -1,11 +1,12 @@
 package ua.bank.bankservice.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import ua.bank.bankservice.model.Account;
 import ua.bank.bankservice.model.Transaction;
 
 public interface TransactionService {
-    void transfer(Account fromAccount, Account toAccount, int amount);
+    void transfer(Account fromAccount, Account toAccount, BigDecimal amount);
 
     List<Transaction> getAllByAccount(int page, int size, Account account);
 }
